S2.3: relocate host behavior out of ergo-adapter
Pre-authorized by freeze-v1.md §4.3.

- moved BufferingRuntimeInvoker, ContextStore, handler coverage, and effect-handler support into ergo-host::host
- updated host imports and tests; no RuntimeHandle::run or RunResult seam change in this commit
- refreshed the S2.3 planning artifact and the v1 boundary/freeze docs for the relocated paths

Plan: docs/ledger/decisions/s2.3-plan.md
