--- crates/kernel/adapter/src/host/mod.rs
@@ -1,9 +1,0 @@
-mod buffering_invoker;
-mod context_store;
-mod coverage;
-mod effects;
-
-pub use buffering_invoker::BufferingRuntimeInvoker;
-pub use context_store::ContextStore;
-pub use coverage::{ensure_handler_coverage, HandlerCoverageError};
-pub use effects::{AppliedWrite, EffectApplyError, EffectHandler, SetContextHandler};
